+++ src/main/java/org/opensearch/knn/index/mapper/KNNVectorFieldMapper.java
@@ -93,6 +93,7 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
      */
     public static class Builder extends ParametrizedFieldMapper.Builder {
         protected Boolean ignoreMalformed;
+        protected final boolean isDerivedSourceEnabled;
 
         protected final Parameter<Boolean> stored = Parameter.storeParam(m -> toType(m).stored, false);
         protected final Parameter<Boolean> hasDocValues = Parameter.docValuesParam(m -> toType(m).hasDocValues, true);
@@ -200,13 +201,15 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
             ModelDao modelDao,
             Version indexCreatedVersion,
             KNNMethodConfigContext knnMethodConfigContext,
-            OriginalMappingParameters originalParameters
+            OriginalMappingParameters originalParameters,
+            boolean isDerivedSourceEnabled
         ) {
             super(name);
             this.modelDao = modelDao;
             this.indexCreatedVersion = indexCreatedVersion;
             this.knnMethodConfigContext = knnMethodConfigContext;
             this.originalParameters = originalParameters;
+            this.isDerivedSourceEnabled = isDerivedSourceEnabled;
         }
 
         @Override
@@ -258,7 +261,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                     modelDao,
                     indexCreatedVersion,
                     originalParameters,
-                    knnMethodConfigContext
+                    knnMethodConfigContext,
+                    isDerivedSourceEnabled
                 );
             }
 
@@ -280,7 +284,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                     ignoreMalformed,
                     stored.get(),
                     hasDocValues.get(),
-                    originalParameters
+                    originalParameters,
+                    isDerivedSourceEnabled
                 );
             }
 
@@ -301,7 +306,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                     metaValue,
                     knnMethodConfigContext,
                     createLuceneFieldMapperInput,
-                    originalParameters
+                    originalParameters,
+                    isDerivedSourceEnabled
                 );
             }
 
@@ -315,7 +321,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                 ignoreMalformed,
                 stored.getValue(),
                 hasDocValues.getValue(),
-                originalParameters
+                originalParameters,
+                isDerivedSourceEnabled
             );
         }
 
@@ -363,7 +370,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                 modelDaoSupplier.get(),
                 parserContext.indexVersionCreated(),
                 null,
-                null
+                null,
+                KNNSettings.isKNNDerivedSourceEnabled(parserContext.getSettings())
             );
             builder.parse(name, parserContext, node);
             builder.setOriginalParameters(new OriginalMappingParameters(builder));
@@ -569,6 +577,7 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
     // values of KNN engine Algorithms hyperparameters.
     protected Version indexCreatedVersion;
     protected Explicit<Boolean> ignoreMalformed;
+    protected final boolean isDerivedSourceEnabled;
     protected boolean stored;
     protected boolean hasDocValues;
     protected VectorDataType vectorDataType;
@@ -589,7 +598,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
         boolean stored,
         boolean hasDocValues,
         Version indexCreatedVersion,
-        OriginalMappingParameters originalMappingParameters
+        OriginalMappingParameters originalMappingParameters,
+        boolean isDerivedSourceEnabled
     ) {
         super(simpleName, mappedFieldType, multiFields, copyTo);
         this.ignoreMalformed = ignoreMalformed;
@@ -599,6 +609,7 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
         updateEngineStats();
         this.indexCreatedVersion = indexCreatedVersion;
         this.originalMappingParameters = originalMappingParameters;
+        this.isDerivedSourceEnabled = isDerivedSourceEnabled;
     }
 
     public KNNVectorFieldMapper clone() {
@@ -831,7 +842,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
             modelDao,
             indexCreatedVersion,
             knnMethodConfigContext,
-            originalMappingParameters
+            originalMappingParameters,
+            isDerivedSourceEnabled
         ).init(this);
     }